runtime: cleanup timer codes
The existing implementation has a poor structure and is a mess.
I just moved exist codes to seprated files and structures
to improve code and logic readability.
No logic has been changed in this cleanup.

Updates #53953
